Clean up formatting
