+++ src/Razor/benchmarks/Microsoft.AspNetCore.Razor.Microbenchmarks/LanguageServer/RazorCSharpFormattingBenchmark.cs
@@ -30,7 +30,7 @@ public class RazorCSharpFormattingBenchmark : RazorLanguageServerBenchmarkBase
 {
     private string _filePath;
 
-    private RazorFormattingService RazorFormattingService { get; set; }
+    private IRazorFormattingService RazorFormattingService { get; set; }
 
     private Uri DocumentUri { get; set; }
 
@@ -143,6 +143,6 @@ public class RazorCSharpFormattingBenchmark : RazorLanguageServerBenchmarkBase
     private void EnsureServicesInitialized()
     {
         var languageServer = RazorLanguageServer.GetInnerLanguageServerForTesting();
-        RazorFormattingService = languageServer.GetRequiredService<RazorFormattingService>();
+        RazorFormattingService = languageServer.GetRequiredService<IRazorFormattingService>();
     }
 }
