+++ src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Debugging/ValidateBreakpointRangeEndpointTest.cs
@@ -136,10 +136,6 @@ public class ValidateBreakpointRangeEndpointTest : SingleServerDelegatingEndpoin
     private RazorRequestContext CreateValidateBreakpointRangeRequestContext(VersionedDocumentContext documentContext)
     {
         var lspServices = new Mock<ILspServices>(MockBehavior.Strict);
-        //lspServices
-        //    .Setup(l => l.GetRequiredService<AdhocWorkspaceFactory>()).Returns(TestAdhocWorkspaceFactory.Instance);
-        //lspServices
-        //    .Setup(l => l.GetRequiredService<RazorFormattingService>()).Returns(TestRazorFormattingService.Instance);
 
         var requestContext = CreateRazorRequestContext(documentContext, lspServices: lspServices.Object);
