+++ src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Formatting/TestRazorFormattingService.cs
@@ -15,13 +15,9 @@ using Moq;
 
 namespace Microsoft.AspNetCore.Razor.LanguageServer.Formatting;
 
-internal class TestRazorFormattingService
+internal static class TestRazorFormattingService
 {
-    private TestRazorFormattingService()
-    {
-    }
-
-    public static async Task<RazorFormattingService> CreateWithFullSupportAsync(
+    public static async Task<IRazorFormattingService> CreateWithFullSupportAsync(
         RazorCodeDocument? codeDocument = null,
         IDocumentSnapshot? documentSnapshot = null,
         ILoggerFactory? loggerFactory = null,
@@ -52,26 +48,25 @@ internal class TestRazorFormattingService
 
         var optionsMonitorCache = new OptionsCache<RazorLSPOptions>();
 
-
-        var optionsMoniter = new TestRazorLSPOptionsMonitor(
+        var optionsMonitor = new TestRazorLSPOptionsMonitor(
             configurationSyncService.Object,
             optionsMonitorCache);
 
         if (razorLSPOptions is not null)
         {
-            await optionsMoniter.UpdateAsync(CancellationToken.None);
+            await optionsMonitor.UpdateAsync(CancellationToken.None);
         }
 
         var passes = new List<IFormattingPass>()
         {
-            new HtmlFormattingPass(mappingService, client, versionCache, optionsMoniter, loggerFactory),
+            new HtmlFormattingPass(mappingService, client, versionCache, optionsMonitor, loggerFactory),
             new CSharpFormattingPass(mappingService, client, loggerFactory),
-            new CSharpOnTypeFormattingPass(mappingService, client, optionsMoniter, loggerFactory),
+            new CSharpOnTypeFormattingPass(mappingService, client, optionsMonitor, loggerFactory),
             new RazorFormattingPass(mappingService, client, loggerFactory),
             new FormattingDiagnosticValidationPass(mappingService, client, loggerFactory),
             new FormattingContentValidationPass(mappingService, client, loggerFactory),
         };
 
-        return new DefaultRazorFormattingService(passes, loggerFactory, TestAdhocWorkspaceFactory.Instance);
+        return new RazorFormattingService(passes, TestAdhocWorkspaceFactory.Instance);
     }
 }
